% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate-cubist.R
\name{hai_auto_cubist}
\alias{hai_auto_cubist}
\title{Boilerplate Workflow}
\usage{
hai_auto_cubist(
  .data,
  .rec_obj,
  .splits_obj = NULL,
  .rsamp_obj = NULL,
  .tune = TRUE,
  .grid_size = 10,
  .num_cores = 1,
  .best_metric = "rmse"
)
}
\arguments{
\item{.data}{The data being passed to the function. The time-series object.}

\item{.rec_obj}{This is the recipe object you want to use. You can use
\code{hai_cubist_data_prepper()} an automatic recipe_object.}

\item{.splits_obj}{NULL is the default, when NULL then one will be created.}

\item{.rsamp_obj}{NULL is the default, when NULL then one will be created. It
will default to creating an \code{\link[rsample:mc_cv]{rsample::mc_cv()}} object.}

\item{.tune}{Default is TRUE, this will create a tuning grid and tuned workflow}

\item{.grid_size}{Default is 10}

\item{.num_cores}{Default is 1}

\item{.best_metric}{Default is "rmse". The only \code{.model_type} you can use with
\code{Cubist} is \code{regression} so use \code{\link[=hai_default_regression_metric_set]{hai_default_regression_metric_set()}}
to get the available metrics. Because of this the \code{.model_type} parameter is
omitted from this function.}
}
\value{
A list
}
\description{
This is a boilerplate function to create automatically the following:
\itemize{
\item recipe
\item model specification
\item workflow
\item tuned model (grid ect)
}
}
\details{
This uses the \code{parsnip::cubist_rules()} with the \code{engine} set to \code{cubist}
}
\examples{
\dontrun{
data <- mtcars

rec_obj <- hai_cubist_data_prepper(data, mpg ~ .)

auto_cube <- hai_auto_cubist(
  .data = data,
  .rec_obj = rec_obj,
  .best_metric = "rmse"
)

auto_cube$recipe_info
}

}
\seealso{
Other Boiler_Plate: 
\code{\link{hai_auto_c50}()},
\code{\link{hai_auto_earth}()},
\code{\link{hai_auto_glmnet}()},
\code{\link{hai_auto_knn}()},
\code{\link{hai_auto_svm_poly}()},
\code{\link{hai_auto_wflw_metrics}()}

Other cubist: 
\code{\link{hai_cubist_data_prepper}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Boiler_Plate}
\concept{cubist}
