% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-kmeans.R
\name{hai_auto_kmeans}
\alias{hai_auto_kmeans}
\title{Automatic K-Means H2O}
\usage{
hai_auto_kmeans(.data)
}
\arguments{
\item{.data}{The data that is to be passed for clustering.}

\item{.split_ratio}{The ratio for training and testing splits.}

\item{.seed}{The default is 1234, but can be set to any integer.}

\item{.centers}{The default is 1. Specify the number of clusters (groups of data)
in a data set.}

\item{.num_folds}{The number of folds for cross-validation}

\item{.max_iterations}{The default is 100. This specifies the number of training
iterations}

\item{.stardardize}{The default is set to TRUE. When TRUE all numeric columns
will be set to zero mean and unit variance.}

\item{.categorical_encoding}{Can be one of the following:
\itemize{
\item "auto"
\item "enum"
\item "one_hot_explicit"
\item "binary"
\item "eigen"
\item "label_encoder"
\item "sort_by_response"
\item "enum_limited"
}}

\item{.initialization_mode}{This can be one of the following:
\itemize{
\item "Random"
\item "Furthest (default)
\item "PlusPlus"
}}

\item{.predictors}{This must be in the form of c("column_1", "column_2", ... "column_n")}
}
\value{

}
\description{
This is a wrapper around the \code{\link[=h2o.kmeans]{h2o.kmeans()}} function that will return a list
object with a lot of useful and easy to use tidy style information.
}
\details{

}
\examples{

}
\seealso{
Other Clustering: 
\code{\link{hai_kmeans_mapped_tbl}()},
\code{\link{hai_kmeans_obj}()},
\code{\link{hai_kmeans_scree_data_tbl}()},
\code{\link{hai_kmeans_scree_plt}()},
\code{\link{hai_kmeans_tidy_tbl}()},
\code{\link{hai_kmeans_user_item_tbl}()}

Other Kmeans: 
\code{\link{hai_kmeans_mapped_tbl}()},
\code{\link{hai_kmeans_obj}()},
\code{\link{hai_kmeans_scree_data_tbl}()},
\code{\link{hai_kmeans_scree_plt}()},
\code{\link{hai_kmeans_tidy_tbl}()},
\code{\link{hai_kmeans_user_item_tbl}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Clustering}
\concept{Kmeans}
