% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-proc-data.R
\name{hai_data_preprocessor}
\alias{hai_data_preprocessor}
\title{Data Preprocessor}
\usage{
hai_data_preprocessor(.data)
}
\arguments{
\item{.data}{The data that you want to process}
}
\value{
A processed data.frame/tibble.
}
\description{
Takes in a data.frame/tibble and transforms it into a processed data.frame/tibble.
}
\details{
This function will get your data ready for processing with many types of ml/ai
models. This is an R translation from the PyCaret pre-processor.
}
\examples{
library(healthyR.data)
library(dplyr)

data_tbl <- healthyR_data\%>\%
   filter(ip_op_flag == "I") \%>\%
   filter(payer_grouping != "Medicare B") \%>\%
   filter(payer_grouping != "?") \%>\%
   select(service_line, payer_grouping) \%>\%
   mutate(record = 1) \%>\%
   as_tibble()

hai_data_preprocessor()

}
\seealso{
\url{https://pycaret.org/preprocessing/}

Other Data Wrangling: 
\code{\link{get_juiced_data}()},
\code{\link{pca_your_recipe}()}

Other Preprocessor: 
\code{\link{hai_data_impute}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Data Wrangling}
\concept{Preprocessor}
