% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-hai-fourier.R
\name{hai_fourier_vec}
\alias{hai_fourier_vec}
\title{Vector Function Fourier}
\usage{
hai_fourier_vec(.x, .period, .order, .scale_type = c("sin", "cos", "sincos"))
}
\arguments{
\item{.x}{A numeric vector}

\item{.period}{The number of observations that complete a cycle}

\item{.order}{The fourier term order}

\item{.scale_type}{A character of one of the following: "sin","cos","sincos"}
}
\value{
A numeric vector
}
\description{
Takes a numeric vector and will return a vector of one of the following:
\itemize{
\item "sin"
\item "cos"
\item "sincos" This will do value = sin(x) * cos(x)
}
}
\details{
Takes a numeric vector and will return a vector of one of the following:
\itemize{
\item "sin"
\item "cos"
\item "sincos"
}

The internal caluclation is straightforward:
\itemize{
\item \code{sin = sin(2 * pi * h * x)}, where \code{h = .order/.period}
\item \code{cos = cos(2 * pi * h * x)}, where \code{h = .order/.period}
\item \code{sincos = sin(2 * pi * h * x) * cos(2 * pi * h * x)} where \code{h = .order/.period}
}

This function can be used on it's own. It is also the basis for the function
\code{\link[=hai_fourier_augment]{hai_fourier_augment()}}.
}
\examples{
suppressPackageStartupMessages(library(dplyr))

len_out    = 25
by_unit    = "month"
start_date = as.Date("2021-01-01")

data_tbl <- tibble(
  date_col = seq.Date(from = start_date, length.out = len_out, by = by_unit),
  a    = rnorm(len_out),
  b    = runif(len_out)
)

vec_1 <- hai_fourier_vec(data_tbl$b, .period = 12, .order = 1, .scale_type = "sin")
vec_2 <- hai_fourier_vec(data_tbl$b, .period = 12, .order = 1, .scale_type = "cos")
vec_3 <- hai_fourier_vec(data_tbl$date_col, .period = 12, .order = 1, .scale_type = "sincos")

plot(data_tbl$b)
lines(vec_1, col = "blue")
lines(vec_2, col = "red")
lines(vec_3, col = "green")

}
\seealso{
Other Vector Function: 
\code{\link{hai_fourier_discrete_vec}()},
\code{\link{hai_hyperbolic_vec}()},
\code{\link{hai_kurtosis_vec}()},
\code{\link{hai_scale_zero_one_vec}()},
\code{\link{hai_skewness_vec}()},
\code{\link{hai_winsorized_move_vec}()},
\code{\link{hai_winsorized_truncate_vec}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Vector Function}
