% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-hai-winsorized-move.R
\name{hai_winsorized_move_vec}
\alias{hai_winsorized_move_vec}
\title{Vector Function Winsorize Move}
\usage{
hai_winsorized_move_vec(.x, .multiple = 3)
}
\arguments{
\item{.x}{A numeric vector}

\item{.multiple}{A positive number indicating how many times the the zero center
mean absolute deviation should be multiplied by for the scaling parameter.}
}
\value{
A numeric vector
}
\description{
Takes a numeric vector and will return a vector of winsorized values.
}
\details{
Takes a numeric vector and will return a winsorized vector of values that have
been moved some multiple from the mean absolute deviation zero center of some
vector. The intent of winsorization is to limit the effect of extreme values.
}
\examples{
suppressPackageStartupMessages(library(dplyr))

len_out    = 25
by_unit    = "month"
start_date = as.Date("2021-01-01")

data_tbl <- tibble(
  date_col = seq.Date(from = start_date, length.out = len_out, by = by_unit),
  a    = rnorm(len_out),
  b    = runif(len_out)
)

vec_1 <- hai_winsorized_move_vec(data_tbl$a, .multiple = 1)

plot(data_tbl$a)
lines(data_tbl$a)
lines(vec_1, col = "blue")

}
\seealso{
\url{https://en.wikipedia.org/wiki/Winsorizing}

This function can be used on it's own. It is also the basis for the function
\code{\link[=hai_winsorized_move_augment]{hai_winsorized_move_augment()}}.

Other Vector Function: 
\code{\link{hai_fourier_discrete_vec}()},
\code{\link{hai_fourier_vec}()},
\code{\link{hai_hyperbolic_vec}()},
\code{\link{hai_scale_zero_one_vec}()},
\code{\link{hai_winsorized_truncate_vec}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Vector Function}
