% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-juiced-data.R
\name{get_juiced_data}
\alias{get_juiced_data}
\title{Get the Juiced Data}
\usage{
get_juiced_data(.recipe_object)
}
\arguments{
\item{.recipe_object}{The recipe you want to pass.}
}
\value{

}
\description{
This is a simple function that will get the juiced data from a recipe.
}
\details{
Instead of typing out something like:
\code{recipe_object \%>\% prep() \%>\% juice() \%>\% glimpse()}
}
\examples{

}
\author{
Steven P. Sanderson II, MPH
}
\concept{Data Wrangling}
