% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hai-impute-missing.R
\name{hai_data_impute}
\alias{hai_data_impute}
\title{Data Preprocessor}
\usage{
hai_data_impute(.recipe_object)
}
\arguments{
\item{.recipe_object}{The data that you want to process}
}
\value{
A processed data.frame/tibble.
}
\description{
Takes in a recipe and will impute missing values using a selected recipe.
To call the recipe use a quoted argument like "median" or "bagged". This function
is not exported but may be called via the ::: method.
}
\details{
This function will get your data ready for processing with many types of ml/ai
models.

This is intended to be used inside of the \code{\link[=hai_data_preprocessor]{hai_data_preprocessor()}} and
therefore is an internal function. This documentation exists to explain the process
and help the user understand the parameters that can be set in the preprocessor function.

The first function used is \code{\link[recipes:step_impute_bag]{recipes::step_impute_bag()}}

Main Recipes Imputation Section:
}
\examples{
library(healthyR.data)
library(dplyr)

data_tbl <- healthyR_data\%>\%
   filter(ip_op_flag == "I") \%>\%
   filter(payer_grouping != "Medicare B") \%>\%
   filter(payer_grouping != "?") \%>\%
   select(service_line, payer_grouping) \%>\%
   mutate(record = 1) \%>\%
   as_tibble()

hai_data_preprocessor()

}
\seealso{
\url{https://recipes.tidymodels.org/reference/index.html#section-step-functions-imputation/}

step_impute_bag

\url{https://recipes.tidymodels.org/reference/step_impute_bag.html}

Other Data Recipes: 
\code{\link{pca_your_recipe}()}

Other Preprocessor: 
\code{\link{hai_data_preprocessor}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Data Recipes}
\concept{Imputation}
\concept{Preprocessor}
\keyword{internal}
