% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hai-impute-missing.R
\name{hai_data_impute}
\alias{hai_data_impute}
\title{Data Preprocessor - Imputation}
\usage{
hai_data_impute(.recipe_object)
}
\arguments{
\item{.recipe_object}{The data that you want to process}

\item{...}{One or more selector functions to choose variables to be imputed.
When used with imp_vars, these dots indicate which variables are used to
predict the missing data in each variable. See selections() for more details}

\item{.impute_vars_with}{A call to imp_vars to specify which variables are
used to impute the variables that can include specific variable names
separated by commas or different selectors (see selections()). If a column is
included in both lists to be imputed and to be an imputation predictor,
it will be removed from the latter and not used to impute itself.}

\item{.seed_value}{To make results reproducible, set the seed.}

\item{.number_of_trees}{This is used for the \code{\link[recipes:step_impute_bag]{recipes::step_impute_bag()}} trees
parameter. This should be an integer.}

\item{.type_of_imputation}{This is a quoted argument and can be one of the following:
\itemize{
\item "bagged"
\item "knn"
\item "linear"
\item "lower"
\item "mean"
\item "median"
\item "mode"
\item "roll"
}}
}
\value{
A processed data.frame/tibble.
}
\description{
Takes in a recipe and will impute missing values using a selected recipe.
To call the recipe use a quoted argument like "median" or "bagged". This function
is not exported but may be called via the ::: method.
}
\details{
This function will get your data ready for processing with many types of ml/ai
models.

This is intended to be used inside of the \code{\link[=hai_data_preprocessor]{hai_data_preprocessor()}} and
therefore is an internal function. This documentation exists to explain the process
and help the user understand the parameters that can be set in the preprocessor function.

The first function used is \code{\link[recipes:step_impute_bag]{recipes::step_impute_bag()}}

\strong{Main Recipes Imputation Section:}
}
\examples{
library(healthyR.data)
library(dplyr)

data_tbl <- healthyR_data\%>\%
   filter(ip_op_flag == "I") \%>\%
   filter(payer_grouping != "Medicare B") \%>\%
   filter(payer_grouping != "?") \%>\%
   select(service_line, payer_grouping) \%>\%
   mutate(record = 1) \%>\%
   as_tibble()

hai_data_preprocessor()

}
\seealso{
\url{https://recipes.tidymodels.org/reference/index.html#section-step-functions-imputation/}

step_impute_bag

\url{https://recipes.tidymodels.org/reference/step_impute_bag.html}

Other Data Recipes: 
\code{\link{pca_your_recipe}()}

Other Preprocessor: 
\code{\link{hai_data_preprocessor}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Data Recipes}
\concept{Imputation}
\concept{Preprocessor}
\keyword{internal}
