% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-proc-data.R
\name{hai_data_preprocessor}
\alias{hai_data_preprocessor}
\title{Data Preprocessor}
\usage{
hai_data_preprocessor(.data)
}
\arguments{
\item{.data}{The data that you want to process}
}
\value{
A processed data.frame/tibble.
}
\description{
Takes in a data.frame/tibble and transforms it into a processed data.frame/tibble.
}
\details{
This function will get your data ready for processing with many types of ml/ai
models. This function works with a recipe object.
}
\examples{
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(recipes))

date_seq <- seq.Date(from = as.Date("2013-01-01"), length.out = 100, by = "month")
val_seq  <- rep(c(rnorm(9), NA), times = 10)
df_tbl   <- tibble(
    date_col = date_seq,
    value    = val_seq
)

rec_obj <- recipe(value ~., df_tbl)

hai_data_preprocessor()

}
\seealso{
Other Data Wrangling: 
\code{\link{get_juiced_data}()},
\code{\link{pca_your_recipe}()}

Other Preprocessor: 
\code{\link{hai_data_impute}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Data Wrangling}
\concept{Preprocessor}
