% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-sets.R
\name{hai_default_regression_metric_set}
\alias{hai_default_regression_metric_set}
\title{Metric Set}
\usage{
hai_default_regression_metric_set()
}
\value{
A yardstick metric set tibble
}
\description{
Default regression metric sets from \code{yardstick}
}
\details{
Default regression metric sets from \code{yardstick}
}
\examples{
hai_default_regression_metric_set()

}
\seealso{
Other Default Metric Sets: 
\code{\link{hai_default_classification_metric_set}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Default Metric Sets}
