% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-hai-get-dist-data.R
\name{hai_get_dist_data_tbl}
\alias{hai_get_dist_data_tbl}
\title{Get Distribution Data Helper}
\usage{
hai_get_dist_data_tbl(.data, .unnest = TRUE, .group_data = FALSE)
}
\value{

}
\description{

}
\details{

}
\examples{

}
\author{
Steven P. Sanderson II, MPH
}
