% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hai-hyperbolic-vec.R
\name{hai_hyperbolic_vec}
\alias{hai_hyperbolic_vec}
\title{Vector Function Hyperbolic}
\usage{
hai_hyperbolic_vec(x, .scale_type = c("sin", "cos", "tan"))
}
\arguments{
\item{.scale_type}{A character of one of the following: "sin","cos","tan"}

\item{.x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
Takes a numeric vector and will return a vector of one of the following:
\itemize{
\item "sin"
\item "cos"
\item "tan"
}
}
\details{
Takes a numeric vector and will return a vector of one of the following:
\itemize{
\item "sin"
\item "cos"
\item "tan"
}

This fucntion can be used on it's own. It is also the basis for the function
\code{\link[=hai_hyperbolic_augment]{hai_hyperbolic_augment()}}.
}
\examples{
suppressPackageStartupMessages(library(dplyr))

len_out    = 10
by_unit    = "month"
start_date = as.Date("2021-01-01")

data_tbl <- tibble(
  date_col = seq.Date(from = start_date, length.out = len_out, by = by_unit),
  a    = rnorm(len_out),
  b    = runif(len_out)
)

hai_hyperbolic_vec(data_tbl$a, .scale_type = "sin")

}
\author{
Steven P. Sanderson II, MPH
}
\concept{Vector Function}
