% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette-color-blind.R
\name{hai_scale_color_colorblind}
\alias{hai_scale_color_colorblind}
\title{Provide Colorblind Compliant Colors}
\usage{
hai_scale_color_colorblind(..., theme = "hai")
}
\description{
Provide Colorblind Compliant Colors
}
