% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-data.R
\name{pca_your_recipe}
\alias{pca_your_recipe}
\title{Perform PCA}
\usage{
pca_your_recipe(.recipe_object, .data, .rotation = TRUE, .threshold = 0.75)
}
\arguments{
\item{.recipe_object}{The recipe object you want to pass.}

\item{.data}{The full dataset that is used in the original recipe object passed
into \code{.recipe_object} in order to obtain the baked data of the transform.}

\item{.rotation}{A boolean that defaults to TRUE, should the rotation be returned}

\item{.threshold}{A number between 0 and 1. A fraction of the total variance
that should be covered by the components.}
}
\value{
A list object with several components
}
\description{
This is a simple function that will perform PCA analysis on a passed recipe.
}
\details{
This is a simple wrapper around some recipes functions to perform a PCA on a
given recipe. This function will output a list and return it invisible.
All of the components of the analysis will be returned in a list as their own
object that can be selected individually. A scree plot is also included.
}
\examples{
suppressPackageStartupMessages(library(timetk))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(purrr))
suppressPackageStartupMessages(library(healthyR.data))
suppressPackageStartupMessages(library(rsample))
suppressPackageStartupMessages(library(recipes))

data_tbl <- healthyR_data \%>\%
    select(visit_end_date_time) \%>\%
    summarise_by_time(
        .date_var = visit_end_date_time,
        .by       = "month",
        value     = n()
    ) \%>\%
    set_names("date_col","value") \%>\%
    filter_by_time(
        .date_var = date_col,
        .start_date = "2013",
        .end_date = "2020"
    )

splits <- initial_split(data = data_tbl, prop = 0.8)

rec_obj <- recipe(value ~., training(splits))

output_list <- pca_your_recipe(rec_obj)
output_list$pca_scree_

}
\seealso{
\url{https://recipes.tidymodels.org/reference/step_pca.html}

Other Data Wrangling: 
\code{\link{get_juiced_data}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Data Recipes}
\concept{Data Wrangling}
\concept{Dimension Reduction}
