% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-hai-fourier-discrete.R,
%   R/step-hai-fourier.R, R/step-hai-hyperbolic.R
\name{required_pkgs.step_hai_fourier_discrete}
\alias{required_pkgs.step_hai_fourier_discrete}
\alias{required_pkgs.step_hai_fourier}
\alias{required_pkgs.step_hai_hyperbolic}
\title{Requited Packages}
\usage{
required_pkgs.step_hai_fourier_discrete(x, ...)

required_pkgs.step_hai_fourier(x, ...)

required_pkgs.step_hai_hyperbolic(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector

A character vector

A character vector
}
\description{
Requited Packages

Requited Packages

Requited Packages
}
\keyword{internal}
