% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-hai-fourier-discrete.R,
%   R/step-hai-fourier.R, R/step-hai-hyperbolic.R, R/step-hai-scale-zero-one.R,
%   R/step-hai-scale-zscore.R, R/step-hai-winsorized-move.R,
%   R/step-hai-winsorized-truncate.R
\name{required_pkgs.step_hai_fourier_discrete}
\alias{required_pkgs.step_hai_fourier_discrete}
\alias{required_pkgs.step_hai_fourier}
\alias{required_pkgs.step_hai_hyperbolic}
\alias{required_pkgs.step_hai_scale_zero_one}
\alias{required_pkgs.step_hai_scale_zscore}
\alias{required_pkgs.step_hai_winsorized_move}
\alias{required_pkgs.step_hai_winsorized_truncate}
\title{Requited Packages}
\usage{
required_pkgs.step_hai_fourier_discrete(x, ...)

required_pkgs.step_hai_fourier(x, ...)

required_pkgs.step_hai_hyperbolic(x, ...)

required_pkgs.step_hai_scale_zero_one(x, ...)

required_pkgs.step_hai_scale_zscore(x, ...)

required_pkgs.step_hai_winsorized_move(x, ...)

required_pkgs.step_hai_winsorized_truncate(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector

A character vector

A character vector

A character vector

A character vector

A character vector

A character vector
}
\description{
Requited Packages

Requited Packages

Requited Packages

Required Packages

Required Packages

Requited Packages

Requited Packages
}
\keyword{internal}
